requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
# Frontend — Elise Scout dashboard
streamlit>=1.36.0
# Google Sheets (optional; used only when --sheets flag is passed)
gspread>=6.0.0
google-auth>=2.28.0
# LLM (optional; used only when EMAIL_LLM_PROVIDER env var is set)
openai>=1.30.0
anthropic>=0.25.0
